lambda generation work in progress
